requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
cssselect==1.1.0
fake-useragent==0.1.11
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
importlib-metadata==4.12.0
lxml==4.9.1
numpy==1.23.3
pandas==1.5.0
parse==1.19.0
psycopg2-binary==2.9.3
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-html==0.10.0
six==1.16.0
slackweb==1.0.5
smmap==5.0.0
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.12
w3lib==2.0.1
websockets==10.3
zipp==3.8.1
